Implement more ESP BLE Descriptor support
BLE HID was broken because the descriptor wasn't marked as
readable. This fixes #9430.

Also add read/write support for client. Servers can only set the
initial value. Setting it later will raise a NotImplementedError.

Setting a user_description will also raise a NotImplementedError.
This can help us find cases that use it.
